pyproject.toml
[project]
name = "pygame-pingpong"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pygame-ce>=2.5.3",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.3.5",
    "taskipy>=1.14.1",
]
